MSVC 1911 warning fix
git-svn-id: https://svn.osgeo.org/gdal/trunk@41121 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/mrf/libLERC/Huffman.cpp
@@ -233,7 +233,7 @@ bool Huffman::ReadCodeTable(const Byte** ppByte, size_t& nRemainingBytesInOut)
 
     m_codeTable.resize(size);
     std::fill( m_codeTable.begin(), m_codeTable.end(),
-                std::pair<short, unsigned int>(0, 0) );
+                std::pair<short, unsigned int>(static_cast<short>(0), 0) );
 
     if( GetIndexWrapAround(i0, size) >= size ||
         GetIndexWrapAround(i1 - 1, size) >= size )
